// apps/server-core/src/adapters/analysis-metadata-recalculator/types.ts
/*
 * Copyright (c) 2025.
 * Author Peter Placzek (tada5hi)
 * For the full copyright and license information,
 * view the LICENSE file that was distributed with this source code.
 */

import type { DataSource } from 'typeorm';

export type AnalysisMetadataRecalculatorConfig = {
    env: string;
    skipAnalysisApproval: boolean;
};

export type AnalysisMetadataRecalculatorContext = {
    dataSource: DataSource;
    config: AnalysisMetadataRecalculatorConfig;
};